requirements.txt
Django==1.10.3
pycrypto==2.6.1
